moving function prompts into their own file
